pyproject.toml
[project]
name = "langchain-pet"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.11.0",
    "isort>=7.0.0",
    "langchain>=1.0.6",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.0.2",
    "python-dotenv>=1.2.1",
]
